# src/frontend/sqeez/src/api/api.yaml
openapi: 3.1.1
info:
  title: Sqeez.Api | v1
  version: 1.0.0
servers:
  - url: http://localhost:5000/
paths:
...
  /api/quiz-attempts/start:
    post:
      tags:
        - QuizAttempts
      summary: POST /api/quiz-attempts/start
      description: "POST /api/quiz-attempts/start\r

        Starts a new attempt for the currently authenticated student."
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartQuizAttemptDto"
          text/json:
            schema:
              $ref: "#/components/schemas/StartQuizAttemptDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/StartQuizAttemptDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
  /api/quiz-attempts/{id}/answer:
    post:
      tags:
        - QuizAttempts
      summary: POST /api/quiz-attempts/{id}/answer
      description: "POST /api/quiz-attempts/{id}/answers\r

        Submits or updates an answer for a specific question."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitQuestionResponseDto"
          text/json:
            schema:
              $ref: "#/components/schemas/SubmitQuestionResponseDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/SubmitQuestionResponseDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuestionAnsweredDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionAnsweredDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuestionAnsweredDto"
  /api/quiz-attempts/{id}/next-question:
    get:
      tags:
        - QuizAttempts
      summary: GET /api/quiz-attempts/{id}/next-question
      description: "GET /api/quiz-attempts/{id}/next-question\r

        Recovers the ID of the next unanswered question and current answered
        count if the frontend loses context."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/NextQuestionProgressDto"
            application/json:
              schema:
                $ref: "#/components/schemas/NextQuestionProgressDto"
            text/json:
              schema:
                $ref: "#/components/schemas/NextQuestionProgressDto"
  /api/quiz-attempts/{id}/complete:
    post:
      tags:
        - QuizAttempts
      summary: POST /api/quiz-attempts/{id}/complete
      description: "POST /api/quiz-attempts/{id}/complete\r

        Locks the attempt and calculates the final score."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDto"
  /api/quiz-attempts/{id}:
    get:
      tags:
        - QuizAttempts
      summary: GET /api/quiz-attempts/{id}
      description: "GET /api/quiz-attempts/{id}\r

        Gets full attempt details. Students see their own attempts, subject
        teachers see attempts for their quizzes,\r

        and admins can see any attempt."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuizAttemptDetailDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDetailDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuizAttemptDetailDto"
    delete:
      tags:
        - QuizAttempts
      summary: DELETE /api/quiz-attempts/{id}
      description: "DELETE /api/quiz-attempts/{id}\r

        Deletes a specific quiz attempt (Allows a teacher to reset a student's
        try)."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
  /api/quiz-attempts/quiz/{quizId}:
    get:
      tags:
        - QuizAttempts
      summary: GET /api/quiz-attempts/quiz/{quizId}
      description: "GET /api/quiz-attempts/quiz/{quizId}\r

        Gets a paginated list of all attempts for a specific quiz."
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        - name: pageNumber
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
            default: 20
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/PagedResponseOfQuizAttemptDto"
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfQuizAttemptDto"
            text/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfQuizAttemptDto"
  /api/quiz-attempts/responses/{responseId}/grade:
    patch:
      tags:
        - QuizAttempts
      summary: PATCH /api/quiz-attempts/responses/{responseId}/grade
      description: "PATCH /api/quiz-attempts/responses/{responseId}/grade\r

        Allows a teacher to manually grade a free-text answer and \"Like\" it."
      parameters:
        - name: responseId
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeQuestionResponseDto"
          text/json:
            schema:
              $ref: "#/components/schemas/GradeQuestionResponseDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/GradeQuestionResponseDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/QuestionResponseDto"
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResponseDto"
            text/json:
              schema:
                $ref: "#/components/schemas/QuestionResponseDto"
  /api/quiz-attempts/{quizId}/attempts:
    delete:
      tags:
        - QuizAttempts
      summary: DELETE /api/quiz-attempts/{quizId}/attempts
      description: "DELETE /api/quizzes/{quizId}/attempts\r

        Deletes all quiz attempts and student responses for a specific quiz."
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
...
components:
  schemas:
...
    AttemptStatus:
      enum:
        - Created
        - Started
        - Completed
        - PendingCorrection
        - Abandoned
      description: Lifecycle state of a quiz attempt.
...
    GradeQuestionResponseDto:
      type: object
      properties:
        score:
          maximum: 1000
          minimum: -1000
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        isLiked:
          type: boolean
          description: Teacher feedback flag for highlighting an answer.
      description: Teacher/admin request for manually grading a free-text response.
...
    NextQuestionProgressDto:
      required:
        - nextQuestionId
        - answeredQuestionsCount
      type: object
      properties:
        nextQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          description: Next unanswered question id, or null when all questions are answered.
          format: int64
        answeredQuestionsCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Number of distinct questions already answered in the attempt.
          format: int32
      description: Quiz attempt navigation state used to recover the next question and
        render progress.
...
    PagedResponseOfQuizAttemptDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/QuizAttemptDto"
          description: Items for the requested page.
        totalCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Total number of items matching the filter before paging is applied.
          format: int32
        pageNumber:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: One-based page number returned by the query.
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Maximum number of items requested for this page.
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Total number of pages calculated from TotalCount and PageSize.
          format: int32
      description: Standard paged response shape used by list endpoints.
...
    QuestionAnsweredDto:
      required:
        - id
        - quizQuestionId
        - responseTimeMs
        - freeTextAnswer
        - isLiked
        - score
        - selectedOptionIds
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        quizQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        responseTimeMs:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        freeTextAnswer:
          type:
            - "null"
            - string
        isLiked:
          type: boolean
        score:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int32
        selectedOptionIds:
          type: array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        correctOptionIds:
          type:
            - "null"
            - array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        correctFreeTextAnswer:
          type:
            - "null"
            - string
        nextQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
      description: Answer-submission response with correctness hints and next-question
        navigation.
    QuestionResponseDto:
      required:
        - id
        - quizQuestionId
        - responseTimeMs
        - freeTextAnswer
        - isLiked
        - score
        - selectedOptionIds
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        quizQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        responseTimeMs:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        freeTextAnswer:
          type:
            - "null"
            - string
        isLiked:
          type: boolean
        score:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int32
        selectedOptionIds:
          type: array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      description: Stored answer response returned in attempt details.
...
    QuizAttemptDetailDto:
      required:
        - id
        - quizId
        - enrollmentId
        - startTime
        - endTime
        - status
        - totalScore
        - mark
        - responses
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        quizId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        enrollmentId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        startTime:
          type:
            - "null"
            - string
          format: date-time
        endTime:
          type:
            - "null"
            - string
          format: date-time
        status:
          $ref: "#/components/schemas/AttemptStatus"
        totalScore:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        mark:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int32
        responses:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResponseDto"
      description: Full quiz attempt view including all submitted responses.
    QuizAttemptDto:
      required:
        - id
        - quizId
        - enrollmentId
        - startTime
        - endTime
        - status
        - totalScore
        - mark
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Attempt identifier.
          format: int64
        quizId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Quiz being attempted.
          format: int64
        enrollmentId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Enrollment that proves the student's access to the quiz subject.
          format: int64
        startTime:
          type:
            - "null"
            - string
          description: UTC timestamp when the attempt started.
          format: date-time
        endTime:
          type:
            - "null"
            - string
          description: UTC timestamp when the attempt was completed or closed.
          format: date-time
        status:
          description: Attempt lifecycle state.
          $ref: "#/components/schemas/AttemptStatus"
        totalScore:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Total score currently awarded for the attempt.
          format: int32
        mark:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          description: Optional derived mark for completed attempts.
          format: int32
        nextQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          description: Next unanswered question id, or null when no pending question
            remains.
          format: int64
        earnedBadges:
          type:
            - "null"
            - array
          items:
            $ref: "#/components/schemas/StudentBadgeBasicDto"
          description: Badges awarded while completing the attempt.
        studentName:
          type:
            - "null"
            - string
          description: Student display name, included only for privileged quiz-owner views.
        studentId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          description: Student id, included only for privileged quiz-owner views.
          format: int64
      description: Quiz attempt summary returned when an attempt is started,
        completed, or listed.
...
    StartQuizAttemptDto:
      required:
        - quizId
        - enrollmentId
      type: object
      properties:
        quizId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        enrollmentId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
      description: Request for starting a quiz attempt for a student's enrollment.
...
    StudentBadgeBasicDto:
      type: object
      properties:
        badgeId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Awarded badge id.
          format: int64
        name:
          type: string
          description: Badge display name.
        iconUrl:
          type:
            - "null"
            - string
          description: Public icon URL, or null when no icon is assigned.
        earnedAt:
          type: string
          description: UTC timestamp when the student earned the badge.
          format: date-time
      description: Compact badge award view embedded in user and quiz attempt responses.
...
    SubmitQuestionResponseDto:
      type: object
      properties:
        quizQuestionId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Question being answered. It must belong to the quiz attempt.
          format: int64
        responseTimeMs:
          maximum: 3600000
          minimum: 0
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          description: Time spent on the question in milliseconds.
          format: int64
        freeTextAnswer:
          maxLength: 4000
          minLength: 0
          type:
            - "null"
            - string
          description: Free-text answer for free-text questions.
        selectedOptionIds:
          maxItems: 1000
          type: array
          items:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
          description: Selected option ids for single-choice or multiple-choice questions.
      description: Request for submitting a response to one quiz question.
...